+++ doc/reference/CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - crd: add capability to set extra annotations on PVC
 - crd: capability disable Prometheus related resources like `PodMonitor` to enable `SoftwareFactory` deployment without
   Prometheus Operator installed on the cluster.
+- weeder: add zuul-weeder service to inspect the global config, available at $fqdn/weeder/
 
 ### Changed